requirements.txt
aiofiles==24.1.0
fastapi==0.116.1
Jinja2==3.1.6
python-jose==3.3.0
motor==3.7.1
passlib==1.7.4
Pillow==11.3.0
pydantic==2.11.7
PyJWT==2.10.1
pymongo==4.13.2
python_bcrypt==0.3.2
Requests==2.32.4
starlette==0.47.1
